+++ bricks/basic/src/editors/index.ts
@@ -1,2 +1,9 @@
 import "./eo-button/index.js";
 import "./eo-submit-buttons/index.js";
+import "./eo-text/index.js";
+import "./eo-search-bar/index.js";
+import "./eo-spin/index.js";
+import "./eo-tag/index.js";
+import "./eo-popover/index.js";
+import "./eo-search/index.js";
+import "./eo-category/index.js";