+++ bricks/visual-builder/src/bootstrap.ts
@@ -15,3 +15,5 @@ import "./pre-generated-preview/index.js";
 import "./pre-generated-table-view/index.js";
 import "./pre-generated-container/index.js";
 import "./property-editor/index.js";
+import "./data-providers/check-editor-by-name.js";
+import "./data-providers/get-data-dependency.js";