+++ bricks/visual-builder/src/property-editor/index.spec.tsx
@@ -13,12 +13,14 @@ describe("visual-builder.property-editor", () => {
       "visual-builder.property-editor"
     ) as PropertyEditor;
 
+    element.editorName = "eo-button-editor";
+
     expect(element.shadowRoot).toBeFalsy();
 
     act(() => {
       document.body.appendChild(element);
     });
-    expect(element.innerHTML).toBe("<div>无数据</div>");
+    expect(element.innerHTML).toBe("");
 
     act(() => {
       document.body.removeChild(element);