+++ bricks/ai/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.23.2](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/ai@1.23.1...@next-bricks/ai@1.23.2) (2024-06-21)
+
+**Note:** Version bump only for package @next-bricks/ai
+
+
+
+
+
+## [1.23.1](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/ai@1.23.0...@next-bricks/ai@1.23.1) (2024-06-20)
+
+**Note:** Version bump only for package @next-bricks/ai
+
+
+
+
+
+# [1.23.0](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/ai@1.22.0...@next-bricks/ai@1.23.0) (2024-06-18)
+
+
+### Features
+
+* **chat-view:** add inputToolbarBrick ([73714da](https://github.com/easyops-cn/next-bricks/commit/73714daae619f8752ca396beeeea6f22713f287a))
+
+
+
+
+
+# [1.22.0](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/ai@1.21.7...@next-bricks/ai@1.22.0) (2024-06-18)
+
+
+### Features
+
+* **chat-view:**  add robotId ([5907886](https://github.com/easyops-cn/next-bricks/commit/590788631dda83d9cfa2f84cd73c5e898b010669))
+
+
+
+
+
+## [1.21.7](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/ai@1.21.6...@next-bricks/ai@1.21.7) (2024-06-17)
+
+**Note:** Version bump only for package @next-bricks/ai
+
+
+
+
+
 ## [1.21.6](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/ai@1.21.5...@next-bricks/ai@1.21.6) (2024-06-11)