+++ bricks/form-platform/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.6](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/form-platform@0.4.5...@next-bricks/form-platform@0.4.6) (2024-06-21)
+
+**Note:** Version bump only for package @next-bricks/form-platform
+
+
+
+
+
+## [0.4.5](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/form-platform@0.4.4...@next-bricks/form-platform@0.4.5) (2024-06-20)
+
+**Note:** Version bump only for package @next-bricks/form-platform
+
+
+
+
+
+## [0.4.4](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/form-platform@0.4.3...@next-bricks/form-platform@0.4.4) (2024-06-17)
+
+**Note:** Version bump only for package @next-bricks/form-platform
+
+
+
+
+
 ## [0.4.3](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/form-platform@0.4.2...@next-bricks/form-platform@0.4.3) (2024-06-11)