+++ bricks/form/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.6](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/form@1.10.5...@next-bricks/form@1.10.6) (2024-06-21)
+
+**Note:** Version bump only for package @next-bricks/form
+
+
+
+
+
+## [1.10.5](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/form@1.10.4...@next-bricks/form@1.10.5) (2024-06-20)
+
+
+### Bug Fixes
+
+* 修改eo-search分类 ([7c3ec49](https://github.com/easyops-cn/next-bricks/commit/7c3ec491b9cf1d501409f664b8694fc08457ec0a))
+
+
+
+
+
+## [1.10.4](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/form@1.10.3...@next-bricks/form@1.10.4) (2024-06-17)
+
+**Note:** Version bump only for package @next-bricks/form
+
+
+
+
+
 ## [1.10.3](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/form@1.10.2...@next-bricks/form@1.10.3) (2024-06-11)