+++ bricks/illustrations/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.24](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/illustrations@1.0.23...@next-bricks/illustrations@1.0.24) (2024-06-21)
+
+**Note:** Version bump only for package @next-bricks/illustrations
+
+
+
+
+
+## [1.0.23](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/illustrations@1.0.22...@next-bricks/illustrations@1.0.23) (2024-06-20)
+
+**Note:** Version bump only for package @next-bricks/illustrations
+
+
+
+
+
+## [1.0.22](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/illustrations@1.0.21...@next-bricks/illustrations@1.0.22) (2024-06-17)
+
+**Note:** Version bump only for package @next-bricks/illustrations
+
+
+
+
+
 ## [1.0.21](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/illustrations@1.0.20...@next-bricks/illustrations@1.0.21) (2024-06-11)