+++ bricks/recent-history/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.32](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/recent-history@1.0.31...@next-bricks/recent-history@1.0.32) (2024-06-21)
+
+**Note:** Version bump only for package @next-bricks/recent-history
+
+
+
+
+
+## [1.0.31](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/recent-history@1.0.30...@next-bricks/recent-history@1.0.31) (2024-06-20)
+
+**Note:** Version bump only for package @next-bricks/recent-history
+
+
+
+
+
+## [1.0.30](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/recent-history@1.0.29...@next-bricks/recent-history@1.0.30) (2024-06-17)
+
+**Note:** Version bump only for package @next-bricks/recent-history
+
+
+
+
+
 ## [1.0.29](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/recent-history@1.0.28...@next-bricks/recent-history@1.0.29) (2024-06-11)